Extract common `ExecutionFailure.Failure` implementation superclass
To handle `assertHasFirstCause` uniformly.

Also, report the _no causes_ failure separately from the _unexpected first
cause_ one.
