+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/executer/OutputScrapingExecutionFailure.java
@@ -303,35 +303,9 @@ public class OutputScrapingExecutionFailure extends OutputScrapingExecutionResul
         }
     }
 
-    private static class Problem implements Failure {
-        final String description;
-        final List<String> causes;
-
+    private static class Problem extends AbstractFailure {
         private Problem(String description, List<String> causes) {
-            this.description = description;
-            this.causes = causes;
-        }
-
-        @Override
-        public void assertHasCause(String message) {
-            if (!causes.contains(message)) {
-                throw new AssertionFailedError(String.format("Expected cause '%s' not found in %s", message, causes));
-            }
-        }
-
-        @Override
-        public void assertHasFirstCause(String message) {
-            String firstCause = causes.stream().findFirst().orElse("");
-            if (!firstCause.equals(message)) {
-                throw new AssertionFailedError(String.format("Expected first cause '%s', got '%s'", message, firstCause));
-            }
-        }
-
-        @Override
-        public void assertHasCauses(int count) {
-            if (causes.size() != count) {
-                throw new AssertionFailedError(String.format("Expecting %d cause(s), got %d", count, causes.size()));
-            }
+            super(description, causes);
         }
     }
 }
